Update Active_Record_Named_Scopes.yml
Adding trailblazer-finder is a Finder Object, that is part of Trailblazer framework. But it is designed to be used on its own as a separate gem.  It was influenced by popular Ransack gem, but in addition to active-record, it can be used with DataMapper or Sequel. It also integrates with Kaminari or Will Paginate, as well as FriendlyId