Add minitest-substitute to mocking.yml
Simple Minitest helper to replace values such as an instance variable of an object or an environment variable for the duration of a block or a group of tests. This comes in very handy when you have to derive from default configuration in order to test some aspects of your code.